backend/requirements.txt
# Python dependencies for FastAPI backend
fastapi
uvicorn
alembic
sqlalchemy
psycopg2-binary
python-jose[cryptography]
argon2-cffi
python-multipart
email-validator
python-dotenv
stripe
pytest
pytest-asyncio
httpx
